# backend/swagger-spec.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}}}},"/auth/signup/student":{"post":{"operationId":"AuthController_addStudent","summary":"Sign up new student","parameters":[],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"firstName":{"type":"string","example":"Eiei"},"lastName":{"type":"string","example":"Eiei"},"email":{"type":"string","example":"eiei@gmail.com"},"phone":{"type":"string","example":"09123456789"},"username":{"type":"string","example":"eiei"},"password":{"type":"string","example":"Eiei123456"},"gender":{"type":"string","example":"m"},"image":{"type":"string","format":"binary"},"preferSubject":{"type":"array","items":{"type":"string","example":"6204f74398648fc94382135f"}}}}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Student"}}}}},"tags":["auth"]}},"/auth/signup/tutor":{"post":{"operationId":"AuthController_addTutor","summary":"Sign up new tutor","parameters":[],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"firstName":{"type":"string","example":"Phu"},"lastName":{"type":"string","example":"Jong"},"email":{"type":"string","example":"phu@jong.com"},"phone":{"type":"string","example":"0123456789"},"username":{"type":"string","example":"phu.jong"},"password":{"type":"string","example":"Phugod123456"},"gender":{"type":"string","example":"m"},"teachSubject":{"type":"array","items":{"type":"string","example":"6204f74398648fc94382135f"}},"priceMin":{"type":"number","example":"10"},"priceMax":{"type":"number","example":"100"},"dutyTime":{"type":"array","items":{"type":"object","properties":{"start":{"type":"string","example":"2022-02-16T08:00:00.000+00:00"},"end":{"type":"string","example":"2022-02-16T09:30:00.000+00:00"}}}},"image":{"type":"string","format":"binary"}}}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Tutor"}}}}},"tags":["auth"]}},"/auth/signin":{"post":{"operationId":"AuthController_signin","summary":"Sign in user to the system with user type","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","properties":{"username":{"type":"string","example":"t eiei"},"password":{"type":"string","example":"Eiei123456"}}}}}},"responses":{"201":{"description":""}},"tags":["auth"]}},"/subject/getById":{"get":{"operationId":"SubjectController_getSubject","summary":"Return a subject whose id is matched","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"properties":{"subjectId":{"type":"string","example":"6204f74398648fc94382135f"}}}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Subject"}}}}},"tags":["subject"]}},"/subject/getAllSubjects":{"get":{"operationId":"SubjectController_getAllSubjects","summary":"Return all subjects","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Subject"}}}}}},"tags":["subject"]}},"/subject/getAllSubjectsName":{"get":{"operationId":"SubjectController_getAllSubjectsName","summary":"Return a list of title of subjects ","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{"subjects":{"type":"array","items":{"type":"string","example":"PAT1"}}}}}}}},"tags":["subject"]}},"/subject/getByLevel":{"get":{"operationId":"SubjectController_getSubjectsByLevel","summary":"Return a list of subjects whose level is matched","parameters":[{"name":"level","required":true,"in":"query","example":"highschool","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{"subjects":{"type":"array","items":{"type":"string","example":"PAT1"}}}}}}}},"tags":["subject"]}},"/subject/getAllLevels":{"get":{"operationId":"SubjectController_getAllLevels","summary":"Return a list of levels","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{"levels":{"type":"array","items":{"type":"string","example":"highschool"}}}}}}}},"tags":["subject"]}},"/subject/getAllSubjectsLevel":{"get":{"operationId":"SubjectController_getAllSubjectsLevel","summary":"Return all subjects with available levels","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{"subject":{"type":"array","items":{"type":"object","properties":{"level":{"type":"string","example":"9subjects"},"id":{"type":"string","example":"6204f74398648fc94382135f"}}}}}}}}}},"tags":["subject"]}},"/subject/create":{"post":{"operationId":"SubjectController_addSubject","summary":"Create new subject","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"properties":{"title":{"type":"string","example":"PAT1"},"level":{"type":"string","example":"highschool"},"description":{"type":"string","example":"the mathematics entrances exam"},"maxScore":{"type":"number","example":"300"}}}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Subject"}}}},"201":{"description":""}},"tags":["subject"]}},"/score/create":{"post":{"operationId":"ScoreController_addScore","summary":"create new score","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"properties":{"tutorId":{"type":"string","example":"621c818daefa29db6f3e806f"},"subjectId":{"type":"string","example":"62072bfd96edc906154249b1"},"score":{"type":"number","example":300},"maxScore":{"type":"number","example":300},"year":{"type":"number","example":2022},"image":{"type":"string","format":"binary"}}}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{"id":{"type":"string","example":"6236aba956e2c3c18ccb0eed"}}}}}}},"tags":["score"],"security":[{"bearer":[]}]}},"/score/delete":{"delete":{"operationId":"ScoreController_deleteScore","summary":"Delete score by tutorId and subjectId","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"properties":{"tutorId":{"type":"string","example":"621c818daefa29db6f3e806f"},"subjectId":{"type":"string","example":"62072bfd96edc906154249b1"}}}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{"id":{"type":"string","example":"6236aba956e2c3c18ccb0eed"}}}}}}},"tags":["score"],"security":[{"bearer":[]}]}},"/score/edit":{"patch":{"operationId":"ScoreController_editScore","summary":"Update score","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Score"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{"scoreId":{"type":"string","example":"6236aba956e2c3c18ccb0eed"}}}}}}},"tags":["score"],"security":[{"bearer":[]}]}},"/score":{"get":{"operationId":"ScoreController_getScore","summary":"Get score by tutorId and subjectId","parameters":[{"name":"tutorId","required":true,"in":"query","example":"621c818daefa29db6f3e806f","schema":{"type":"string"}},{"name":"subjectId","required":true,"in":"query","example":"6236aba956e2c3c18ccb0eed","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{"subject":{"type":"string","example":"123456"},"score":{"type":"number","example":300},"maxScore":{"type":"number","example":300},"scoreImage":{"type":"string","example":"https://skyandtelescope.org/wp-content/uploads/Earth-from-Mars_m.jpg"}}}}}}},"tags":["score"],"security":[{"bearer":[]}]}},"/score/scores/{tutorId}":{"get":{"operationId":"ScoreController_getAllScore","summary":"Get all scores by tutorId","parameters":[{"name":"tutorId","required":true,"in":"query","example":"621c818daefa29db6f3e806f","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Score"}}}}},"tags":["score"],"security":[{"bearer":[]}]}},"/score/getAllPendingScore":{"get":{"operationId":"ScoreController_getAllPendingScore","summary":"Get all pending scores","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Score"}}}}}},"tags":["score"],"security":[{"bearer":[]}]}},"/score/{scoreId}":{"get":{"operationId":"ScoreController_getScoreById","summary":"Get score by scoreId ","parameters":[{"name":"scoreId","required":true,"in":"path","example":"6236aba956e2c3c18ccb0eed","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Score"}}}}},"tags":["score"],"security":[{"bearer":[]}]}},"/score/{tutorId}/{subjectId}/approve":{"patch":{"operationId":"ScoreController_approveScore","summary":"Approve score which has tutorId and subjectId","parameters":[{"name":"tutorId","required":true,"in":"path","schema":{"type":"string"}},{"name":"subjectId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Score"}}}}},"tags":["score"],"security":[{"bearer":[]}]}},"/score/{tutorId}/{subjectId}/reject":{"patch":{"operationId":"ScoreController_rejectScore","summary":"Reject score which has tutorId and subjectId","parameters":[{"name":"tutorId","required":true,"in":"path","schema":{"type":"string"}},{"name":"subjectId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Score"}}}}},"tags":["score"],"security":[{"bearer":[]}]}},"/student/recommend":{"post":{"operationId":"StudentController_recommendTutor","summary":"Recommend tutors based on student's prefered subjects","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","properties":{"studentId":{"type":"string","example":"621c8c3d363377298c2bf8b2"}}}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Tutor"}}}}}},"tags":["student"],"security":[{"bearer":[]},{"bearer":[]}]}},"/student/getById":{"get":{"operationId":"StudentController_getStudentById","summary":"Return a student whose id is matched","parameters":[{"name":"id","required":true,"in":"query","example":"621c8c3d363377298c2bf8b2","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Student"}}}}},"tags":["student"],"security":[{"bearer":[]}]}},"/student/{id}":{"patch":{"operationId":"StudentController_updateStudent","summary":"Update student's account information","parameters":[{"name":"id","required":true,"in":"path","example":"621c8c3d363377298c2bf8b2","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"firstName":{"type":"string","example":"Phu"},"lastName":{"type":"string","example":"Jong"},"email":{"type":"string","example":"phu@jong.com"},"image":{"type":"string","format":"binary"},"phone":{"type":"string","example":"0123456789"},"username":{"type":"string","example":"phu.jong"},"gender":{"type":"string","example":"m"},"preferSubject":{"type":"array","items":{"type":"string","example":"6204f74398648fc94382135f"}}}}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Student"}}}}},"tags":["student"],"security":[{"bearer":[]}]}},"/tutor/search":{"post":{"operationId":"TutorController_searchTutor","summary":"Search tutor by name or subject","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"properties":{"text":{"type":"string","example":"text to search"}}}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Tutor"}}}}}},"tags":["tutor"],"security":[{"bearer":[]}]}},"/tutor/match":{"post":{"operationId":"TutorController_matchTutor","summary":"Match tutors based on student's matching options","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"properties":{"subjectName":{"type":"string","example":"mathematics"},"level":{"type":"string","example":"highschool"},"priceMin":{"type":"number","example":"100"},"priceMax":{"type":"number","example":"200"},"availabilityStudent":{"type":"array","items":{"type":"object","properties":{"availabilityDate":{"type":"string","example":"2020-01-31"},"availabilityTimeFrom":{"type":"string","example":"10:00"},"availabilityTimeTo":{"type":"string","example":"12:00"}}}}}}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Tutor"}}}}}},"tags":["tutor"],"security":[{"bearer":[]}]}},"/tutor/{id}":{"patch":{"operationId":"TutorController_updateTutor","summary":"Update tutor's account information","parameters":[{"name":"id","required":true,"in":"path","example":"621c818daefa29db6f3e806f","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","properties":{"firstName":{"type":"string","example":"Poom"},"lastName":{"type":"string","example":"Jong"},"email":{"type":"string","example":"phu@jong.com"},"image":{"type":"string","format":"binary"},"phone":{"type":"string","example":"0123456789"},"username":{"type":"string","example":"phu.jong"},"gender":{"type":"string","example":"m"},"avgRating":{"type":"number","example":"5.0"},"successMatch":{"type":"number","example":"10"},"teachSubject":{"type":"array","items":{"type":"string","example":"6204f74398648fc94382135f"}},"priceMin":{"type":"number","example":"10"},"priceMax":{"type":"number","example":"100"},"dutyTime":{"type":"array","items":{"type":"object","properties":{"start":{"type":"string","example":"2022-02-16T08:00:00.000+00:00"},"end":{"type":"string","example":"2022-02-16T09:30:00.000+00:00"}}}}}}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Tutor"}}}}},"tags":["tutor"],"security":[{"bearer":[]}]}},"/tutor/getById":{"get":{"operationId":"TutorController_getTutorById","summary":"Return a tutor whose id is matched","parameters":[{"name":"id","required":true,"in":"query","example":"621c818daefa29db6f3e806f","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Tutor"}}}}},"tags":["tutor"],"security":[{"bearer":[]}]}},"/tutor/score":{"get":{"operationId":"TutorController_getTutorScore","summary":"Return all scores of tutor whose id is matched","parameters":[{"name":"id","required":true,"in":"query","example":"621c818daefa29db6f3e806f","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Score"}}}}}},"tags":["tutor"],"security":[{"bearer":[]}]}},"/chat/request":{"post":{"operationId":"ChatController_addChat","summary":"Get chat of tutor and student or create new chatroom if not existing","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","properties":{"tutorId":{"type":"string","example":"621c818daefa29db6f3e806f"},"studentId":{"type":"string","example":"621c8c3d363377298c2bf8b2"}}}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Chat"}}}}},"tags":["chat"],"security":[{"bearer":[]}]}},"/chat/delete":{"delete":{"operationId":"ChatController_deleteChat","summary":"Delete chat between tutor and student","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","properties":{"tutorId":{"type":"string","example":"621c818daefa29db6f3e806f"},"studentId":{"type":"string","example":"621c8c3d363377298c2bf8b2"}}}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{"deleteChat":{"type":"string","example":"6239a861391503e1556c0b1e"}}}}}}},"tags":["chat"],"security":[{"bearer":[]}]}},"/chat":{"get":{"operationId":"ChatController_getChat","summary":"Get chat by tutor id and student id","parameters":[{"name":"tutorId","required":true,"in":"query","example":"621c818daefa29db6f3e806f","schema":{"type":"string"}},{"name":"studentId","required":true,"in":"query","example":"621c8c3d363377298c2bf8b2","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Chat"}}}}},"tags":["chat"],"security":[{"bearer":[]}]}},"/chat/getChatsTutor":{"get":{"operationId":"ChatController_getChatsTutor","summary":"Get all chats of tutor whose id is matched","parameters":[{"name":"tutorId","required":true,"in":"query","example":"621c818daefa29db6f3e806f","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Chat"}}}}}},"tags":["chat"],"security":[{"bearer":[]}]}},"/chat/getChatsStudent":{"get":{"operationId":"ChatController_getChatsStudent","summary":"Get all chats of student whose id is matched","parameters":[{"name":"studentId","required":true,"in":"query","example":"621c8c3d363377298c2bf8b2","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Chat"}}}}}},"tags":["chat"],"security":[{"bearer":[]}]}},"/chat/tutorAcceptChat/{chatId}":{"post":{"operationId":"ChatController_tutorAcceptChat","summary":"Tutor accept requested chat","parameters":[{"name":"chatId","required":true,"in":"path","example":"6239a861391503e1556c0b1e","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Chat"}}}}},"tags":["chat"],"security":[{"bearer":[]}]}},"/chat/declineChat/{chatId}":{"post":{"operationId":"ChatController_declineChat","summary":"Tutor decline requested chat","parameters":[{"name":"chatId","required":true,"in":"path","example":"6239a861391503e1556c0b1e","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{"chatId":{"type":"string","example":"6239a861391503e1556c0b1e"}}}}}}},"tags":["chat"],"security":[{"bearer":[]}]}},"/appointment":{"post":{"operationId":"AppointmentController_createAppointment","summary":"Create an appointment with a tutor and a student","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","properties":{"chatId":{"type":"string","example":"6239a861391503e1556c0b1e"},"subjectId":{"type":"string","example":"6204f74398648fc94382135f"},"price":{"type":"number","example":"191"},"startTime":{"type":"string","example":"Sat Apr 09 2022 22:50:22 GMT+0800 (Indochina Time)"},"endTime":{"type":"string","example":"Sat Apr 09 2022 22:50:22 GMT+0100 (Indochina Time)"}}}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Tutor"}}}}},"tags":["appointment"],"security":[{"bearer":[]}]}},"/appointment/{userType}/{id}/{status}":{"get":{"operationId":"AppointmentController_getAppointmentsByType","summary":"Get list of appointments of user with user type filtered by status","parameters":[{"name":"userType","required":true,"in":"path","example":"tutor","schema":{"type":"string"}},{"name":"id","required":true,"in":"path","example":"621c818daefa29db6f3e806f","schema":{"type":"string"}},{"name":"status","required":true,"in":"path","example":"confirmed","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Appointment"}}}}}},"tags":["appointment"],"security":[{"bearer":[]}]}},"/appointment/chat/{chatId}":{"get":{"operationId":"AppointmentController_getAppointmentsByChat","summary":"Get list of appointments by chat id","parameters":[{"name":"chatId","required":true,"in":"path","example":"6239a861391503e1556c0b1e","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Appointment"}}}}}},"tags":["appointment"],"security":[{"bearer":[]}]}},"/appointment/{userType}/{id}":{"get":{"operationId":"AppointmentController_getAppointments","summary":"Get list of appointments of user with user type","parameters":[{"name":"userType","required":true,"in":"path","example":"tutor","schema":{"type":"string"}},{"name":"id","required":true,"in":"path","example":"621c818daefa29db6f3e806f","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Appointment"}}}}}},"tags":["appointment"],"security":[{"bearer":[]}]}},"/appointment/{appointmentId}":{"delete":{"operationId":"AppointmentController_cancelAppointment","summary":"Cancel an appointment that matched with the id","parameters":[{"name":"appointmentId","required":true,"in":"path","example":"621c818daefa29db6f3e806f","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Appointment"}}}}}},"tags":["appointment"],"security":[{"bearer":[]}]}},"/appointment/student/accept/{appointmentId}":{"patch":{"operationId":"AppointmentController_studentAcceptAppointment","summary":"Accept an appointment that matched with the id","parameters":[{"name":"appointmentId","required":true,"in":"path","example":"621c818daefa29db6f3e806f","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["appointment"],"security":[{"bearer":[]}]}}},"info":{"title":"Tuture","description":"A platform to match tutor and student.","version":"1.0","contact":{}},"tags":[{"name":"appointment","description":""},{"name":"auth","description":""},{"name":"chat","description":""},{"name":"review","description":""},{"name":"score","description":""},{"name":"student","description":""},{"name":"subject","description":""},{"name":"tutor","description":""}],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"Student":{"type":"object","properties":{"_id":{"type":"string"},"firstName":{"type":"string"},"lastName":{"type":"string"},"email":{"type":"string"},"phone":{"type":"string"},"username":{"type":"string"},"gender":{"type":"string"},"profileUrl":{"type":"string"},"preferSubject":{"type":"array","items":{"type":"string"}}},"required":["_id","firstName","lastName","email","phone","username","gender","profileUrl","preferSubject"]},"Tutor":{"type":"object","properties":{"_id":{"type":"string"},"firstName":{"type":"string"},"lastName":{"type":"string"},"email":{"type":"string"},"phone":{"type":"string"},"username":{"type":"string"},"gender":{"type":"string"},"profileUrl":{"type":"string"},"avgRating":{"type":"number"},"successMatch":{"type":"number"},"teachSubject":{"type":"array","items":{"type":"string"}},"priceMax":{"type":"number"},"priceMin":{"type":"number"},"dutyTime":{"type":"array","items":{"type":"string"}}},"required":["_id","firstName","lastName","email","phone","username","gender","profileUrl","avgRating","successMatch","teachSubject","priceMax","priceMin","dutyTime"]},"Subject":{"type":"object","properties":{"_id":{"type":"string"},"title":{"type":"string"},"level":{"type":"string"},"description":{"type":"string"},"maxScore":{"type":"number"}},"required":["_id","title","level","description","maxScore"]},"Score":{"type":"object","properties":{"_id":{"type":"string"},"tutorId":{"type":"string"},"subjectId":{"type":"string"},"currentScore":{"type":"number"},"imageUrl":{"type":"string"},"maxScore":{"type":"number"},"year":{"type":"number"},"status":{"type":"string"},"validator":{"type":"string"}},"required":["_id","tutorId","subjectId","currentScore","imageUrl","maxScore","year","status","validator"]},"Chat":{"type":"object","properties":{"_id":{"type":"string"},"tutorId":{"type":"string"},"studentId":{"type":"string"},"messages":{"type":"array","items":{"type":"string"}},"accepted":{"type":"boolean"}},"required":["_id","tutorId","studentId","messages","accepted"]},"Appointment":{"type":"object","properties":{"_id":{"type":"string"},"tutorId":{"type":"string"},"studentId":{"type":"string"},"status":{"type":"string"},"price":{"type":"number"},"subjectId":{"type":"string"},"startTime":{"format":"date-time","type":"string"},"endTime":{"format":"date-time","type":"string"}},"required":["_id","tutorId","studentId","status","price","subjectId","startTime","endTime"]}}}}